requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
anyio==4.11.0
arabic-reshaper==3.0.0
asgiref==3.10.0
asn1crypto==1.5.1
attrs==25.4.0
billiard==4.2.2
CacheControl==0.14.3
cachetools==6.2.1
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
cryptography==46.0.3
cssselect2==0.8.0
dj-database-url==3.0.1
Django==5.2.7
django-cloudinary-storage==0.3.0
django-crontab==0.7.1
django-environ==0.12.0
django-jazzmin==3.0.1
django-mathfilters==1.0.0
django-phonenumber-field==8.3.0
django-tinymce==5.0.0
django-twilio==0.14.3.3
fcm-django==2.3.1
firebase_admin==7.1.0
flower==2.0.1
freetype-py==2.5.1
frozenlist==1.8.0
google-api-core==2.28.1
google-auth==2.42.1
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
humanize==4.14.0
hyperframe==6.1.0
idna==3.11
kombu==5.5.4
lxml==6.0.2
msgpack==1.1.2
multidict==6.7.0
oscrypto==1.3.0
packaging==25.0
phonenumbers==9.0.17
pillow==12.0.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.28.0
pycparser==2.23
psycopg2-binary==2.9.9
pyfcm==2.1.0
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
PyJWT==2.10.1
pypdf==6.1.3
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
redis==7.0.1
reportlab==4.4.4
requests==2.32.5
rlPyCairo==0.4.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
svglib==1.6.0
swapper==1.4.0
tinycss2==1.4.0
tornado==6.5.2
twilio==9.8.5
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
webencodings==0.5.1
whitenoise==6.11.0
xhtml2pdf==0.2.17
yarl==1.22.0
